Add the brood improvement feature framework + simulated annealing.
The BroodImproverHelper class takes a set of solutions, applies a random brood improvement algorithm
on the set of solutions using multiple threads.
MyPartition is a helper class that partitions a list.
FoodFactory methods have been changed to static.
InitialSolutionsGenerator - added possibility to generate N solutions.
NutrientsIdealValuesHelper's methods are a shared resource, so they are declared synchronized.